update ziggurat, implement normlimit_by_axis.
replace ziggurat.cc with ziggurat.cpp (now version 18 October 2013): this fixes
the r4_* functions, which didn't return random numbers on a 64bit-architecture.
implement normlimit_by_axis in eigenmat.cc, which allows using dropout.

+++ eigenmat/eigenmat.h
@@ -1,3 +1,5 @@
+#include <inttypes.h>
+
 #define ERROR_INCOMPATIBLE_DIMENSIONS -1
 #define CUBLAS_ERROR -2
 #define CUDA_ERROR -3
@@ -16,8 +18,8 @@ struct eigenmat {
 };
 
 struct rnd_struct {
-  unsigned long seed;
-  int kn[128];
+  uint32_t seed;
+  uint32_t kn[128];
   float fn[128], wn[128];
 };
